-- Create shopping list --
Adding the new list Item by clicking on the ' Click me ' button, or by pressing the Enter key.
If the input box is empty then, pressing the Enter key or clicking the add button will give the user an alert message to inform him that the there is nothing inside the input box.