-- Add features to the shopping list --
Add an input box to the HTML file , so the user can add the quantity of the added item, and make the quantity optional.
Beside that we check the loading state of the document,
and we build a function contain our code and use it to start the js code.
And change some selectors in the CSS file.